# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: MI Academy API
  version: "1.0.0"
  description: |
    MI Academy backend API — parent dashboard, child progress, games, and sync.
    All contracts are versioned. Breaking changes increment the major version.
  contact:
    name: MI Academy Platform Team
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Current production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ParentProfileResponse:
      type: object
      required: [id, display_name, language, timezone, has_pin, created_at]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        language:
          type: string
          enum: [vi, en]
        timezone:
          type: string
        has_pin:
          type: boolean
        created_at:
          type: string
          format: date-time
        children:
          type: array
          items:
            $ref: "#/components/schemas/ChildResponse"
    ParentProfileUpdate:
      type: object
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
    SetPinRequest:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          pattern: "^[0-9]{4,6}$"
    VerifyPinRequest:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          pattern: "^[0-9]{4,6}$"
    VerifyPinResponse:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
    ChildResponse:
      type: object
      required: [id, nickname, age_group, preferred_language, created_at]
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
          maxLength: 50
        birth_year:
          type: integer
        age_group:
          $ref: "#/components/schemas/AgeGroup"
        grade_level:
          type: string
        avatar_id:
          type: string
        preferred_language:
          type: string
          enum: [vi, en]
        daily_time_limit_minutes:
          type: integer
        created_at:
          type: string
          format: date-time
...
    AgeGroup:
      type: string
      enum: [junior, explorer, master]
...
paths:
...
  /parent/profile:
    get:
      tags: [parent]
      summary: Get parent profile
      responses:
        "200":
          description: Parent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentProfileResponse"
    put:
      tags: [parent]
      summary: Update parent profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParentProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentProfileResponse"
  /parent/pin:
    post:
      tags: [parent]
      summary: Set parent PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPinRequest"
      responses:
        "200":
          description: PIN set
  /parent/pin/verify:
    post:
      tags: [parent]
      summary: Verify parent PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPinRequest"
      responses:
        "200":
          description: PIN verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPinResponse"
...